GIA-68 Add transaction management
* GIA-68 Add transaction management

* GIA-68 Add transaction management - change type trainigDuration

* GIA-68 Add transaction management - add package com.gym.crm.dto.request;

* GIA-68 Add transaction management - Introduce TransactionScope for unified transaction handling

* GIA-68 Add transaction management - change type in test for trainingDuration

* GIA-68 Add transaction management - Formatting correction

+++ src/main/java/com/gym/crm/dto/request/CreateTrainingRequest.java
@@ -8,6 +8,7 @@ import jakarta.validation.constraints.Size;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -33,5 +34,5 @@ public class CreateTrainingRequest {
     private final LocalDate trainingDate;
 
     @Positive(message = "Training duration must be a positive number")
-    private final int trainingDuration;
+    private final BigDecimal trainingDuration;
 }
